Fix unit tests and address comments
